misc: factor subtitle similarity helpers
