+++ misc/string_similarity.h
@@ -1,6 +1,6 @@
+#pragma once
+
 /*
- * Language code utility functions
- *
  * This file is part of mpv.
  *
  * mpv is free software; you can redistribute it and/or
@@ -17,18 +17,8 @@
  * License along with mpv.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef MP_LANGUAGE_H
-#define MP_LANGUAGE_H
+#include <stddef.h>
 
-#include <stdbool.h>
-
-#include "misc/bstr.h"
-#include "common/common.h"
-
-// Result numerically higher => better match. 0 == no match.
-int mp_match_lang(char **langs, const char *lang);
-char **mp_get_user_langs(void);
-bstr mp_guess_lang_from_filename(bstr name, int *lang_start, enum track_flags *flags);
-bool mp_language_is_suffix_token(const char *token);
-
-#endif /* MP_LANGUAGE_H */
+char *mp_normalize_base_name(void *ta_ctx, const char *path);
+double mp_similarity_ratio(const char *a, const char *b);
+int mp_levenshtein_dist(const char *a, const char *b);
